Add a lock to avoid multiple workers to emit at the same time.
Closes #5442

(cherry picked from commit 02bb455315a17a39cb0d936a12a13fbbdb887bb3)
