+++ CHANGES/5442.bugfix
@@ -1,0 +1,1 @@
+Added a lock to avoid multiple workers sending metrics at the same time.
